==> p06.Rmd <==
---
title: "Portfolio 6"
---

> The results of my first year project indicates that blocking acetylcholine receptiors in the nucleus accumbens doesn't affect impulsivity. However, previous studies have shown that 24 hours after the injecting scopolomine (Ach muscarinic receptor antagonist) the PPE mRNA less. This is the mRNA that produces PPE, which a precursor of a mu-opioid neurotransmitter. We have found that stimulitaing mu-opioid receptors in NAc would increase impulsivity. So one of the reasons that we didn't find significant result is that it takes time for Ach anatagnist to work with the opioid system to influence impulsivity. On that note, it would be helpful to analyze rats' implulsivity the day after we give the injection to see if scopolomine produces any effect that we didn't after in the 1-hr experiment right after the injection. We happen to have the data, so this project aims to analyze it. 

### Prep.
```{r lib}
library(readxl)
library(tidyverse)
library(haven)
library(lme4)
library(lmerTest)
library(ez)
library(emmeans)
library(ggpattern)
library(afex)
```

```{r read-in-data}
IM_DD1_2day <- read_excel("data/fyp/1dayafterinjection/IM-3-21-Basic.xlsx")
IM_DD2_2day <- read_excel("data/fyp/1dayafterinjection/IM-3-24-Basic.xlsx")
IM_DD3_2day <- read_excel("data/fyp/1dayafterinjection/IM-3-27-Basic.xlsx")
IM_DD4_2day <- read_excel("data/fyp/1dayafterinjection/IM-4-02-Basic.xlsx")
IM_DD5_2day <- read_excel("data/fyp/1dayafterinjection/IM-4-06-Basic.xlsx")
IM_all_clean <- read_rds("data/fyp/IM_all_clean.rds")
```

```{r combine-cal-efficiency}
IM_all_2day <- bind_rows(IM_DD1_2day, IM_DD2_2day, IM_DD3_2day, IM_DD4_2day, IM_DD5_2day)
IM_all_2day$Efficiency <- IM_all_2day$ReinforcersEarned / IM_all_2day$RightLeverPress * 100
IM_all_2day <- IM_all_2day %>%
  filter(Subject != "IM10")
```

### 1. Analysis & Visualization

#### 1.1 Look at the data.
```{r vis}
IM_all_2day$Group <- factor(IM_all_2day$Group, levels = c("M_HIGH", "M_LOW", "VEH", "S_LOW", "S_HIGH"))

#Lever press plot
IM_all_2day %>%
  ggplot(aes(x = Group, y = RightLeverPress)) +
  geom_boxplot() +
  geom_point(aes(color = Subject)) +  
  stat_summary(fun = mean, geom = "line", group = 1, color = "black") +  
  stat_summary(fun = mean, geom = "point", color = "black") 

#Reinforcers earned plot
IM_all_2day %>%
  ggplot(aes(x = Group, y = ReinforcersEarned)) +
  geom_boxplot() +
  geom_point(aes(color = Subject)) +  
  stat_summary(fun = mean, geom = "line", group = 1, color = "black") +  
  stat_summary(fun = mean, geom = "point", color = "black") 

#Efficiency plot
IM_all_2day %>%
  ggplot(aes(x = Group, y = Efficiency)) +
  geom_boxplot() +
  geom_point(aes(color = Subject)) +  
  stat_summary(fun = mean, geom = "line", group = 1, color = "black") +  
  stat_summary(fun = mean, geom = "point", color = "black")  
```
This is fucking beautiful. This is exactly we predicted!

#### 1.2 Now some ANOVAs.
Right Lever Press:
```{r anova-lever-press}
RLP_results <- ezANOVA(
  data = IM_all_2day,
  dv = .(RightLeverPress),     # Dependent variable
  wid = .(Subject),  # Within-subject variable (ID)
  within = .(Group)   # Repeated measures factor
)
RLP_results #sig
#post hoc
pairwise.t.test(IM_all_2day$RightLeverPress, IM_all_2day$Group, paired = TRUE, p.adjust.method = "bonferroni")
#S_HIGH pressed significantly less levers than all other groups
```

Reinforcers Earned:
```{r anova-reinforcers-earned}
RLP_results <- ezANOVA(
  data = IM_all_2day,
  dv = .(ReinforcersEarned),     # Dependent variable
  wid = .(Subject),  # Within-subject variable (ID)
  within = .(Group)   # Repeated measures factor
)
RLP_results #sig
#post hoc
pairwise.t.test(IM_all_2day$ReinforcersEarned, IM_all_2day$Group, paired = TRUE, p.adjust.method = "bonferroni")
#S_HIGH pressed significantly less levers than all other groups
```

Efficiency:
```{r anova-efficiency}
RLP_results <- ezANOVA(
  data = IM_all_2day,
  dv = .(Efficiency),     # Dependent variable
  wid = .(Subject),  # Within-subject variable (ID)
  within = .(Group)   # Repeated measures factor
)
RLP_results #sig
#post hoc
pairwise.t.test(IM_all_2day$Efficiency, IM_all_2day$Group, paired = TRUE, p.adjust.method = "bonferroni")
#S_HIGH pressed significantly less levers than all other groups
```
They are all significant! So injecting scopolamine does affect rats' impulsivity, but the effect is delayed till the second day!

#### 1.3 Now more bar plots.
Right Lever Press:
```{r bar-plot-lever-press3}
# total level press
# Calculate standard error
means <- IM_all_2day %>%
  group_by(Group) %>%
  summarize(
    MeanScore = mean(RightLeverPress),
    SE = sd(RightLeverPress) / sqrt(n()),
    .groups = "drop"
  )

# Add error bars
ggplot(means, aes(x = Group, y = MeanScore, fill = Group)) +
  geom_bar(stat = "identity", width = 0.7) +
  geom_errorbar(aes(ymin = MeanScore - SE, ymax = MeanScore + SE), width = 0.15, color = "darkgrey") +
  labs(title = "Total Lever Presses 2nd Day", x = NULL, y = NULL) +
  scale_fill_manual(values = c("M_HIGH" = "#397DB0", "M_LOW" = "#9DC7DD", "VEH" = "#fbf0d9", "S_LOW" = "#9ED17B", "S_HIGH" = "#3D9F3C")) +
  scale_x_discrete(labels = c("M_HIGH" = "Mecamylamine\nHigh Dose", "M_LOW" = "Mecamylamine\nLow Dose", "VEH" = "Vehicle", "S_LOW" = "Scopolamine\nLow Dose", "S_HIGH" = "Scopolamine\nHigh Dose")) +
  theme_minimal() +
  theme(legend.position = "none", axis.text.x = element_text(size = 11, face = "bold"), axis.text.y = element_text(size = 15, face = "bold"),  plot.title = element_text(hjust = 0.5, size = 20, face = "bold"))
```

Reinforcers Earned:
```{r bar-plot-reinforcers-earned}
# total level press
# Calculate standard error
means <- IM_all_2day %>%
  group_by(Group) %>%
  summarize(
    MeanScore = mean(ReinforcersEarned),
    SE = sd(ReinforcersEarned) / sqrt(n()),
    .groups = "drop"
  )

# Add error bars
ggplot(means, aes(x = Group, y = MeanScore, fill = Group)) +
  geom_bar(stat = "identity", width = 0.7) +
  geom_errorbar(aes(ymin = MeanScore - SE, ymax = MeanScore + SE), width = 0.15, color = "darkgrey") +
  labs(title = "Reinforcers Earned 2nd Day", x = NULL, y = NULL) +
  scale_fill_manual(values = c("M_HIGH" = "#397DB0", "M_LOW" = "#9DC7DD", "VEH" = "#fbf0d9", "S_LOW" = "#9ED17B", "S_HIGH" = "#3D9F3C")) +
  scale_x_discrete(labels = c("M_HIGH" = "Mecamylamine\nHigh Dose", "M_LOW" = "Mecamylamine\nLow Dose", "VEH" = "Vehicle", "S_LOW" = "Scopolamine\nLow Dose", "S_HIGH" = "Scopolamine\nHigh Dose")) +
  theme_minimal() +
  theme(legend.position = "none", axis.text.x = element_text(size = 11, face = "bold"), axis.text.y = element_text(size = 15, face = "bold"),  plot.title = element_text(hjust = 0.5, size = 20, face = "bold"))
```

Efficiency:
```{r bar-plot-efficiency2}
# total level press
# Calculate standard error
means <- IM_all_2day %>%
  group_by(Group) %>%
  summarize(
    MeanScore = mean(Efficiency),
    SE = sd(Efficiency) / sqrt(n()),
    .groups = "drop"
  )

# Add error bars
ggplot(means, aes(x = Group, y = MeanScore, fill = Group)) +
  geom_bar(stat = "identity", width = 0.7) +
  geom_errorbar(aes(ymin = MeanScore - SE, ymax = MeanScore + SE), width = 0.15, color = "darkgrey") +
  labs(title = "Efficiency 2nd Day", x = NULL, y = NULL) +
  scale_fill_manual(values = c("M_HIGH" = "#397DB0", "M_LOW" = "#9DC7DD", "VEH" = "#fbf0d9", "S_LOW" = "#9ED17B", "S_HIGH" = "#3D9F3C")) +
  scale_x_discrete(labels = c("M_HIGH" = "Mecamylamine\nHigh Dose", "M_LOW" = "Mecamylamine\nLow Dose", "VEH" = "Vehicle", "S_LOW" = "Scopolamine\nLow Dose", "S_HIGH" = "Scopolamine\nHigh Dose")) +
  theme_minimal() +
  theme(legend.position = "none", axis.text.x = element_text(size = 11, face = "bold"), axis.text.y = element_text(size = 15, face = "bold"),  plot.title = element_text(hjust = 0.5, size = 20, face = "bold"))
```

### 2. Combine data for two days.
Since we see a different effect of drug condition for the injection day and one day post-injection. It would be helpful to combine data for both days and visualize them.

First we combine datasets for two days.
```{r combine-two-day-datasets}
IM_all_clean$DayAfterInjection <- 0
IM_all_2day$DayAfterInjection <- 1
IM_all_grand <- bind_rows(IM_all_clean, IM_all_2day)
IM_all_grand$DayAfterInjection <- as.factor(IM_all_grand$DayAfterInjection)
```

Then we need to map out the bar plots, shading dataset from the day after the injection.
Total lever presses:
```{r bar-plot-lever-press2}
# total level press
# Calculate standard error
means <- IM_all_grand %>%
  group_by(Group, DayAfterInjection) %>%
  summarize(
    MeanScore = mean(RightLeverPress),
    SE = sd(RightLeverPress) / sqrt(n()),
    .groups = "drop"
  )

# Create the bar plot with shading
ggplot(means, aes(x = Group, y = MeanScore, fill = Group, pattern = DayAfterInjection)) +
  geom_bar_pattern(
    stat = "identity", 
    position = position_dodge(),  # Adjust position for dodging
    width = 0.7,
    pattern_fill = "black",
    pattern_density = 0.05, 
    pattern_spacing = 0.05, 
    pattern_angle = 45,
    color = "black"  # Add black borders to the bars
  ) +
  geom_errorbar(aes(ymin = MeanScore - SE, ymax = MeanScore + SE), 
                width = 0.15, color = "darkgrey", position = position_dodge(width = 0.7)) +
  labs(title = "Total Lever Presses", x = NULL, y = NULL, pattern = NULL) +
  scale_fill_manual(values = c(
    "M_HIGH" = "#397DB0", 
    "M_LOW" = "#9DC7DD", 
    "VEH" = "#fbf0d9", 
    "S_LOW" = "#9ED17B", 
    "S_HIGH" = "#3D9F3C"
  )) +
  scale_x_discrete(labels = c(
    "M_HIGH" = "Mecamylamine\nHigh Dose", 
    "M_LOW" = "Mecamylamine\nLow Dose", 
    "VEH" = "Vehicle", 
    "S_LOW" = "Scopolamine\nLow Dose", 
    "S_HIGH" = "Scopolamine\nHigh Dose"
  )) +
  scale_pattern_manual(values = c("0" = "none", "1" = "stripe"), 
                       labels = c("0" = "Day of Injection", "1" = "Day After Injection")) + # Add labels to explain patterns
  theme_minimal() +
  theme(
    legend.position = "none",
    axis.text.x = element_text(size = 11, face = "bold"), 
    axis.text.y = element_text(size = 15, face = "bold"),  
    plot.title = element_text(hjust = 0.5, size = 20, face = "bold")
  ) 
```

Total reinforcers earned:
```{r bar-plot-reinforcers-earned2}
# total level press
# Calculate standard error
means <- IM_all_grand %>%
  group_by(Group, DayAfterInjection) %>%
  summarize(
    MeanScore = mean(ReinforcersEarned),
    SE = sd(ReinforcersEarned) / sqrt(n()),
    .groups = "drop"
  )

# Create the bar plot with shading
ggplot(means, aes(x = Group, y = MeanScore, fill = Group, pattern = DayAfterInjection)) +
  geom_bar_pattern(
    stat = "identity", 
    position = position_dodge(),  # Adjust position for dodging
    width = 0.7,
    pattern_fill = "black",
    pattern_density = 0.05, 
    pattern_spacing = 0.05, 
    pattern_angle = 45,
    color = "black"  # Add black borders to the bars
  ) +
  geom_errorbar(aes(ymin = MeanScore - SE, ymax = MeanScore + SE), 
                width = 0.15, color = "darkgrey", position = position_dodge(width = 0.7)) +
  labs(title = "Reinforcers Earned", x = NULL, y = NULL, pattern = NULL) +
  scale_fill_manual(values = c(
    "M_HIGH" = "#397DB0", 
    "M_LOW" = "#9DC7DD", 
    "VEH" = "#fbf0d9", 
    "S_LOW" = "#9ED17B", 
    "S_HIGH" = "#3D9F3C"
  )) +
  scale_x_discrete(labels = c(
    "M_HIGH" = "Mecamylamine\nHigh Dose", 
    "M_LOW" = "Mecamylamine\nLow Dose", 
    "VEH" = "Vehicle", 
    "S_LOW" = "Scopolamine\nLow Dose", 
    "S_HIGH" = "Scopolamine\nHigh Dose"
  )) +
  scale_pattern_manual(values = c("0" = "none", "1" = "stripe"), 
                       labels = c("0" = "Day of Injection", "1" = "Day After Injection")) + # Add labels to explain patterns
  theme_minimal() +
  theme(
    legend.position = "none",
    axis.text.x = element_text(size = 11, face = "bold"), 
    axis.text.y = element_text(size = 15, face = "bold"),  
    plot.title = element_text(hjust = 0.5, size = 20, face = "bold")
  ) 
```

Efficiency:
```{r bar-plot-efficiency}
# Efficiency
# Calculate standard error
means <- IM_all_grand %>%
  group_by(Group, DayAfterInjection) %>%
  summarize(
    MeanScore = mean(Efficiency),
    SE = sd(Efficiency) / sqrt(n()),
    .groups = "drop"
  )

# Create the bar plot with shading
ggplot(means, aes(x = Group, y = MeanScore, fill = Group, pattern = DayAfterInjection)) +
  geom_bar_pattern(
    stat = "identity", 
    position = position_dodge(),  # Adjust position for dodging
    width = 0.7,
    pattern_fill = "black",
    pattern_density = 0.05, 
    pattern_spacing = 0.05, 
    pattern_angle = 45,
    color = "black"  # Add black borders to the bars
  ) +
  geom_errorbar(aes(ymin = MeanScore - SE, ymax = MeanScore + SE), 
                width = 0.15, color = "darkgrey", position = position_dodge(width = 0.7)) +
  labs(title = "Efficiency", x = NULL, y = NULL, pattern = NULL) +
  scale_fill_manual(values = c(
    "M_HIGH" = "#397DB0", 
    "M_LOW" = "#9DC7DD", 
    "VEH" = "#fbf0d9", 
    "S_LOW" = "#9ED17B", 
    "S_HIGH" = "#3D9F3C"
  )) +
  scale_x_discrete(labels = c(
    "M_HIGH" = "Mecamylamine\nHigh Dose", 
    "M_LOW" = "Mecamylamine\nLow Dose", 
    "VEH" = "Vehicle", 
    "S_LOW" = "Scopolamine\nLow Dose", 
    "S_HIGH" = "Scopolamine\nHigh Dose"
  )) +
  scale_pattern_manual(values = c("0" = "none", "1" = "stripe"), 
                       labels = c("0" = "Day of Injection", "1" = "Day After Injection")) + # Add labels to explain patterns
  theme_minimal() +
  theme(
    legend.position = "none",
    axis.text.x = element_text(size = 11, face = "bold"), 
    axis.text.y = element_text(size = 15, face = "bold"),  
    plot.title = element_text(hjust = 0.5, size = 20, face = "bold")
  ) 
```

```{r write-out-data, cache=TRUE}
#long to wide
IM_all_2day_wide <- IM_all_2day %>%
  select(Subject, Group, Efficiency, ReinforcersEarned, RightLeverPress, LeftLeverPress, NosePokes) %>%
  pivot_wider(
    names_from = Group,  # Columns to create
    values_from = c(Efficiency, ReinforcersEarned, RightLeverPress, LeftLeverPress, NosePokes)     # Values to fill
  )

write_sav(IM_all_2day_wide, "data/fyp/IM_all_2day_wide.sav")
```

#### 2.1 Run a 5 (drug) * 2 (day) repeated measures ANOVA

```{r two-way-ANOVA}
# Run the ANOVA
############################### lever press ############################################
result <- aov_ez(
  id = "Subject",                # The subject identifier
  dv = "RightLeverPress",                     # Dependent variable
  data = IM_all_grand,              # Data frame containing your data
  within = c("Group", "DayAfterInjection"),     # Within-subject factors
  type = 3                       # Type III sums of squares (default)
)

# Display the ANOVA results
print(result)

# Post hoc comparisons
posthoc <- emmeans(result, pairwise ~ Group * DayAfterInjection, adjust = "bonferroni")

# Display post hoc results
summary(posthoc)

# Simple effects for Drug at each Day
emmeans(result, ~ Group | DayAfterInjection)

############################### reinforcers earned ###################
result <- aov_ez(
  id = "Subject",                # The subject identifier
  dv = "ReinforcersEarned",                     # Dependent variable
  data = IM_all_grand,              # Data frame containing your data
  within = c("Group", "DayAfterInjection"),     # Within-subject factors
  type = 3                       # Type III sums of squares (default)
)

# Display the ANOVA results
print(result)

# Post hoc comparisons
posthoc <- emmeans(result, pairwise ~ Group * DayAfterInjection, adjust = "bonferroni")

# Display post hoc results
summary(posthoc)

# Simple effects for Drug at each Day
emmeans(result, ~ Group | DayAfterInjection)

############################### efficiency ###################
result <- aov_ez(
  id = "Subject",                # The subject identifier
  dv = "Efficiency",                     # Dependent variable
  data = IM_all_grand,              # Data frame containing your data
  within = c("Group", "DayAfterInjection"),     # Within-subject factors
  type = 3                       # Type III sums of squares (default)
)

# Display the ANOVA results
print(result)

# Post hoc comparisons
posthoc <- emmeans(result, pairwise ~ Group * DayAfterInjection, adjust = "bonferroni")

# Display post hoc results
summary(posthoc)

# Simple effects for Drug at each Day
emmeans(result, ~ Group | DayAfterInjection)
```

### Conclusion

The plots show an interesting trend -- scopolamine potentially decreases rats' efficiency in the DRL-20 task the next day after injection, which means that the effect of blocking Ach muscarinic receptors on rats' impulsivity maybe delayed!
