==> p08.Rmd <==
---
title: "Portfolio 8"
---

> This project continue to analyze my FYP data. It aims to visualize what rats were doing during the one hour session, which helps us to understand what kind of effect drugs has on rats. Are they lever pressing constantly? Are they nose poking? And are they getting sugar pallets?

### 0. Prep.

```{r read-in-data}
library(tidyverse)

IM_Carray_all_clean <- readRDS("~/Desktop/WFU/25Spring/R/portfolio.pdf/data/fyp/IM_Carray_all_clean.rds")
```

Carray is a compound variable where each event time in the box is saved with an event code:
.1 = Right lever press
.2 = Left Lever press
.3 = Nosepoke
.4 = FREE PELLET
.5 = EARNED PELLET
The number BEFORE the decimal is a time stamp. The two rightmost numbers are .XY
For example, 1800 would be 18.00 seconds.

```{r filter}
# We need to filter out all rows that are not lever presses! 
#IM_Carray_all_clean <- IM_Carray_all_clean %>%
#  filter(Event == 1)

IM_Carray_all_clean <- IM_Carray_all_clean %>%
  filter(!(Event %in% c(2, 4)))
```

```{r interval}
#calculate the inter-response time
IM_Carray_all_clean <- IM_Carray_all_clean %>%
  group_by(Subject, Day) %>%
  mutate(Interval = c(0, diff(Time))) %>%
  ungroup()
```

```{r eval=FALSE}
IM_Carray_all_clean$Reward <- ifelse(IM_Carray_all_clean$Interval >= 20, 1, 0)
IM_Carray_all_clean$Dependency <- if(IM_Carray_all_clean$Reward )
  
IM_Carray_all_clean <- IM_Carray_all_clean %>%
  mutate(Dependency = if_else(
    Reward == 0 & lag(Reward, default = NA) %in% c(0, 1),
    0,
    1,
    missing = 0
  ))
```

### 1. Plot the lever press event across the 1-hour session for every rat across all drugs.

```{r plot-each-rat}
Subject <- c("IM1", "IM2", "IM3", "IM4", "IM5", "IM6", "IM7", "IM8", "IM9", "IM11", "IM12")
for (i in Subject) {
plot <- IM_Carray_all_clean %>% 
  filter(Event == 1) %>% 
  filter(Subject == i) %>% 
  ggplot(aes(x = Time, y = Event, color = Drug)) +
  geom_point(position = position_jitterdodge()) +
  facet_wrap(~ Drug, nrow = 5)

print(plot)
}
```

The pattern seems to indicate that high dose of scopolamine makes rats less persistent, like they got frustrated or something and gave up lever pressing.

### 2. Plot lever press event of S_HIGH and VEH for all rats. 

```{r plot-veh-scoph-lever-press}
IM_Carray_all_clean$Subject <- factor(IM_Carray_all_clean$Subject, levels = c("IM1", "IM2", "IM3", "IM4", "IM5", "IM6", "IM7", "IM8", "IM9", "IM11", "IM12"))

IM_Carray_all_clean %>%
  filter(Event == 1) %>% 
  filter(Drug == "S_HIGH") %>%
  ggplot(aes(x = Time, y = Event, color = Subject)) +
  geom_point(position = position_jitterdodge()) +
  facet_wrap(~ Subject, nrow = 11)

IM_Carray_all_clean %>% 
  filter(Event == 1) %>% 
  filter(Drug == "VEH") %>%
  ggplot(aes(x = Time, y = Event, color = Subject)) +
  geom_point(position = position_jitterdodge()) +
  facet_wrap(~ Subject, nrow = 11)
```

### 3. Plot all 3 events (lever press, nose poke, and sugar pallet eared) of S_HIGH and VEH for all rats. 

```{r plot-veh-scoph-nose-poke, fig.height=10}
IM_Carray_all_clean %>%
  filter(Drug == "S_HIGH") %>%
  ggplot(aes(x = Time, y = Event, color = Subject)) +
  geom_point(position = position_jitterdodge()) +
  facet_wrap(~ Subject, nrow = 11) +
  scale_y_discrete(
    breaks = c(1, 3, 5),
    labels = c("lever press", "nose poke", "sugar pallet earned")
  )

IM_Carray_all_clean %>%
  filter(Drug == "VEH") %>%
  ggplot(aes(x = Time, y = Event, color = Subject)) +
  geom_point(position = position_jitterdodge()) +
  facet_wrap(~ Subject, nrow = 11) +
  scale_y_discrete(
    breaks = c(1, 3, 5),
    labels = c("lever press", "nose poke", "sugar pallet earned")
  )
```

The patterns of nose poke and sugar pallet earned are consistent with that of lever press, supporting our decrease appetitive motivation hypothesis (rats who got high dose scopolomine were less interested in lever pressing to get sugar pallets).
