==> p10.Rmd <==
---
title: "Portfolio 10"
---

> This project contintue to analyze my FYP data by examine how the likelihood of a rat’s lever press depends on its previous press using survival analysis and logistic regression. Specifically, it examines whether drug treatments influence the rats’ tendency to wait longer or respond prematurely following a previous rewarded or premature response.


### 0. Prep.

```{r read-in-data}
library(tidyverse)
library(survival)
library(survminer)
library(lme4)
library(ggeffects)

IM_Carray_all_clean <- readRDS("~/Desktop/WFU/25Spring/R/portfolio.pdf/data/fyp/IM_Carray_all_clean.rds")
```

Carray is a compound variable where each event time in the box is saved with an event code:
.1 = Right lever press
.2 = Left Lever press
.3 = Nosepoke
.4 = FREE PELLET
.5 = EARNED PELLET
The number BEFORE the decimal is a time stamp. The two rightmost numbers are .XY
For example, 1800 would be 18.00 seconds.

```{r filter}
# We need to filter out all rows that are not lever presses! 
IM_Carray_all_clean <- IM_Carray_all_clean %>%
  filter(Event == 1)

#IM_Carray_all_clean <- IM_Carray_all_clean %>%
#  filter(!(Event %in% c(2, 4)))
```

```{r interval}
#calculate the inter-response time
IM_Carray_all_clean <- IM_Carray_all_clean %>%
  group_by(Subject, Day) %>%
  mutate(Interval = c(0, diff(Time))) %>%
  ungroup()

# get outcome from Interval
IM_Carray_all_clean <- IM_Carray_all_clean %>%
  arrange(Subject, Drug) %>%
  group_by(Subject) %>%
  mutate(
    Outcome = ifelse(Interval >= 20, "Rewarded", "Premature"),
    PrevOutcome = lag(Outcome)
  ) %>%
  ungroup()
```

## 1. Survival analysis

## 1.1 Survival models

```{r survival}
# Remove the first press for each rat (no previous press to measure survival time from)
surv_data <- IM_Carray_all_clean %>%
  filter(!is.na(Interval) & !is.na(PrevOutcome)) %>%
  filter(Drug %in% c("VEH", "S_HIGH"))

# Create a Surv object: Interval as survival time, all "events" are press (so status = 1)
surv_obj <- Surv(time = surv_data$Interval, event = rep(1, nrow(surv_data)))

# Fit a Cox model using previous outcome as predictor
cox_model_simple <- coxph(surv_obj ~ PrevOutcome, data = surv_data)
summary(cox_model_simple)

cox_model_interact <- coxph(surv_obj ~ PrevOutcome * Drug, data = surv_data)
summary(cox_model_interact)

# Compare models
anova(cox_model_simple, cox_model_interact)
```

Both the main effect of Drug and the comparison of models are not significant, indicating that although the hazard ratio decreased followed by a rewarded response, the effect does not differ by drugs. Interestingly, this is not consistent with the condition probability results I got in portfolio piece 9. I'm not sure if it's because of the model used or that here we using a continuous DV and in PP9 we analyzed a categorical DV.

## 1.2 Plot the survival models

```{r vis}
ggsurvplot(survfit(surv_obj ~ PrevOutcome, data = surv_data),
           pval = TRUE, conf.int = TRUE, xlim = c(0, 60),   
           ggtheme = theme_minimal())

ggsurvplot(survfit(surv_obj ~ interaction(PrevOutcome, Drug), data = surv_data),
           pval = TRUE,  
           xlim = c(0, 60),   
           legend.title = "Condition",
           legend.labs = c("Premature-S_HIGH", "Rewarded-S_HIGH", "Premature-VEH", "Rewarded-VEH"),
           ggtheme = theme_minimal())
```

These plots clearly show that when rats were rewarded in the previous response, they are more likely to wait longer for the next response.

## 2. Logistic Regression 

### 2.1. Model

```{r glm}
# Binary code the outcome: 1 = Rewarded, 0 = Premature
glm_data <- IM_Carray_all_clean %>%
  mutate(
    Rewarded = ifelse(Outcome == "Rewarded", 1, 0),
    PrevRewarded = ifelse(PrevOutcome == "Rewarded", 1,
                          ifelse(PrevOutcome == "Premature", 0, NA))
  ) %>%
  filter(Drug %in% c("VEH", "S_HIGH"))

# random intercept model with PrevRewarded as a predictor
glmm_model <- glmer(Rewarded ~ PrevRewarded + (1 | Subject), data = glm_data, family = "binomial")
summary(glmm_model)

# random slope model with PrevRewarded & Drug as predictors
glmm_model <- glmer(
  Rewarded ~ PrevRewarded * Drug + (Drug | Subject),
  data = glm_data,
  family = binomial
)
summary(glmm_model)
```

From the random slope model predicting rats' lever press success from the previous lever press, we can see that previous lever press has a marginally significant effect on rats' next lever presses, but drug doesn't significantly influence this effect.

### 2.2. Plot

```{r glm_vis}
pred <- ggpredict(glmm_model, terms = c("PrevRewarded", "Drug"))

plot(pred) +
  labs(
    title = "Predicted Probability of Rewarded Response",
    x = "Previous Response Rewarded (0 = No, 1 = Yes)",
    y = "Predicted Probability",
    color = "Drug"
  ) +
  theme_minimal()
```


## Conclusion

Finally, it's interesting to see how different models/methods provide converging/different results.
